Fix a rather important mistake in the db generation script, do not drop the table once all structure has been migrated

+++ lib/dbstruct/SpotStruct_abs.php
@@ -573,7 +573,9 @@ abstract class SpotStruct_abs {
             if ($cacheCount > 500) {
                 throw new SqlErrorException("Unable to upgrade your cache table. Please run the script 'migrate-cache.php' before this.");
             } else {
-                $this->dropTable("cache");
+                if ($this->columnExists('cache', 'content')) {
+                    $this->dropTable("cache");
+                } # if
             } # else
 		} # if
